% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_or.R
\name{calc_or}
\alias{calc_or}
\title{Title: Function to calculate odds ratios given quantiles of PGS distribution}
\usage{
calc_or(pgs, outcome, pgs_quantiles = seq(0, 1, 0.1), create_plot = T)
}
\arguments{
\item{pgs}{A numeric list with the PGS value per individual.}

\item{outcome}{A binary vector encoding the outcome per individual (case = 1 and control = 0).}

\item{pgs_quantiles}{A list specifying all intervals of the PGS quantiles to contrast. Values must span from 0 to 1. Defaults to deciles (0.1 increments).}

\item{create_plot}{A logical indicating whether to generate a plot of the OR point estimates and intervals. Defaults to True.}
}
\value{
Results: A tibble with odds ratios, confidence intervals and p-values for each PGS quantile when compared to the baseline (lowest quantile).

Quantiles: A list of the PGS quantiles used.

Plot: A ggplot object showing odds ratios (y axis) by PGS quantile (x axis).
}
\description{
This function calculates odds ratios of a binary outcome across quantiles of a PGS distribution
}
\examples{
set.seed(123)
n <- 10000
pgs <- rnorm(n, mean = 0, sd = 4)
beta <- 0.3
mu <- qlogis(0.3)
prob_case <- plogis(mu + beta * pgs)
outcome <- rbinom(n, size = 1, prob = prob_case)
calc_or(pgs, outcome)
}
